it's an int64

+++ include/gim/ecs/component_array.hpp
@@ -51,7 +51,7 @@ class ComponentArray : public IComponentArray {
 		// Find the index of the entity.
 		auto it = std::find(entities.begin(), entities.end(), entity);
 		// Get the index of the entity.
-		int index = std::distance(entities.begin(), it);
+		int64_t index = std::distance(entities.begin(), it);
 
 		// Check if the index is valid.
 		if (index >= entities.size()) {
